Small Change to Data and Looking into Hash Changes
Need to figure out why hash changes are throwing so many errors. adding
in the window.location.href variable does change the url properly and
the back/forward buttons behave as they should this way but the console
is showing lots of "max call stack size exceeded" errors. Not sure how
impactful this is.